requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.23,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.12,<2.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
python-jose>=3.3,<4.0
python-multipart>=0.0.9,<0.1
psycopg2-binary>=2.9,<3.0

# ML / CV / OCR
numpy>=1.24,<3.0
opencv-python-headless>=4.8,<5.0
Pillow>=10,<12
pytesseract>=0.3.10,<0.4
torch>=2.2,<3.0
torchvision>=0.17,<1.0
scikit-learn>=1.3,<2.0

# Testing (optional but included for convenience)
pytest>=7,<9
